+++ src/client/components/PaginationMore.tsx
@@ -1,5 +1,6 @@
 import * as React from "react";
 
+// eslint-disable-next-line
 const PaginationMore: React.SFC<{}> = () => (
   <li className="pagination__item__more">
     <i className="material-icons">more_horiz</i>