+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - ESLint Errors
+- Text in "How To Apply" section spilling over the container if a long url was present
 
 ## [0.26.0] - 2020-07-17