+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `<LoadingIndicator />` color better reflects theme
 - The `<Header />` is now clickable, and will send you to the Search page
+- `<JobCard />` won't display alert if image is missing
 
 ### Removed